refactor: remove BaseContainedButton due to complexity
