+++ dev-client/src/components/buttons/DialogButton.tsx
@@ -15,9 +15,11 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-import {TouchableRippleProps} from 'react-native-paper';
+import {useCallback, useState} from 'react';
+import {StyleSheet, Text} from 'react-native';
+import {TouchableRipple, TouchableRippleProps} from 'react-native-paper';
 
-import {BaseContainedButton} from 'terraso-mobile-client/components/Buttons/BaseContainedButton';
+import {convertColorProp} from 'terraso-mobile-client/components/util/nativeBaseAdapters';
 
 export type DialogButtonType = 'default' | 'destructive' | 'outlined';
 
@@ -27,11 +29,111 @@ export type DialogButtonProps = {
   onPress?: TouchableRippleProps['onPress'];
 };
 
-export const DialogButton = ({label, type, onPress}: DialogButtonProps) => (
-  <BaseContainedButton
-    label={label}
-    context="dialog"
-    type={type}
-    onPress={onPress}
-  />
-);
+export const DialogButton = ({
+  label,
+  type = 'default',
+  onPress,
+}: DialogButtonProps) => {
+  const [pressed, setPressed] = useState(false);
+  const onPressIn = useCallback(() => setPressed(true), [setPressed]);
+  const onPressOut = useCallback(() => setPressed(false), [setPressed]);
+
+  const containerStyles = CONTAINER_STYLES[type];
+  const containerStyle = pressed
+    ? containerStyles.pressed
+    : containerStyles.default;
+  const contentStyle = CONTENT_STYLES[type];
+
+  return (
+    <TouchableRipple
+      style={[styles.container, containerStyle]}
+      accessibilityRole="button"
+      accessibilityLabel={label}
+      onPress={onPress}
+      onPressIn={onPressIn}
+      onPressOut={onPressOut}>
+      <>
+        <Text style={[styles.label, contentStyle]}>{label}</Text>
+      </>
+    </TouchableRipple>
+  );
+};
+
+export type BaseContainedButtonProps = {
+  label: string;
+  type?: DialogButtonType;
+  onPress?: TouchableRippleProps['onPress'];
+};
+
+const styles = StyleSheet.create({
+  container: {
+    alignSelf: 'flex-start',
+    borderWidth: 1,
+    borderRadius: 4,
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    paddingVertical: 10,
+    paddingHorizontal: 24,
+  },
+  containerDefault: {
+    borderColor: convertColorProp('primary.main'),
+    backgroundColor: convertColorProp('primary.main'),
+  },
+  containerDefaultPressed: {
+    borderColor: convertColorProp('primary.dark'),
+    backgroundColor: convertColorProp('primary.dark'),
+  },
+  containerDestructive: {
+    borderColor: convertColorProp('error.main'),
+    backgroundColor: convertColorProp('error.main'),
+  },
+  containerDestructivePressed: {
+    borderColor: convertColorProp('error.dark'),
+    backgroundColor: convertColorProp('error.dark'),
+  },
+  containerOutlined: {
+    borderColor: convertColorProp('text.primary'),
+    backgroundColor: convertColorProp('transparent'),
+  },
+  containerOutlinedPressed: {
+    borderColor: convertColorProp('text.primary'),
+    backgroundColor: convertColorProp('action.selected'),
+  },
+  contentDefault: {
+    color: convertColorProp('primary.contrast'),
+  },
+  contentDestructive: {
+    color: convertColorProp('error.contrast'),
+  },
+  contentOutlined: {
+    color: convertColorProp('text.primary'),
+  },
+  label: {
+    fontWeight: '500',
+    textTransform: 'capitalize',
+    fontSize: 14,
+    lineHeight: 20,
+  },
+});
+
+const CONTAINER_STYLES = {
+  default: {
+    default: styles.containerDefault,
+    pressed: styles.containerDefaultPressed,
+  },
+  destructive: {
+    default: styles.containerDestructive,
+    pressed: styles.containerDestructivePressed,
+  },
+  outlined: {
+    default: styles.containerOutlined,
+    pressed: styles.containerOutlinedPressed,
+  },
+};
+
+const CONTENT_STYLES = {
+  default: styles.contentDefault,
+  destructive: styles.contentDestructive,
+  outlined: styles.contentOutlined,
+};
